+++ app/Http/Controllers/Admin/CategoryController2.php
@@ -134,6 +134,14 @@ class CategoryController2 extends CommonController
     //DELETE    删除分类 | admin/Category2/{Category2}
     public function destroy($cate_id){
 
+      $test = Category1::where('cate_pid',$cate_id)->count();
+      if($test){
+        return $data=[
+          'status'=>1,
+          'msg'=>'有子分类 不能删除！',
+        ];
+      }
+
       $result = Category2::where('cate_id',$cate_id)->delete();
       Category2::where('cate_pid',$cate_id)->update(['cate_pid'=>0]);
       if($result){
